requirements.txt
web3==6.5.0
requests==2.31.0
python-dotenv==1.0.0
ipfshttpclient==0.8.0a2
